Addressed requested changes
