+++ tests/integration/test_agentd/test_agentd_multi_server/data/configuration_templates/wazuh_conf.yaml
