+++ tests/integration/test_agentd/test_agentd_parametrized_reconnections/data/configuration_templates/wazuh_conf.yaml
