+++ tests/integration/test_agentd/test_agentd_reconnection/data/configuration_templates/wazuh_conf.yaml
