+++ tests/integration/test_agentd/test_agentd_state/data/configuration_templates/wazuh_conf.yaml
