+++ tests/integration/test_agentd/test_agentd_state/data/test_cases/wazuh_state_config_tests.yaml
