+++ tests/integration/test_agentd/test_agentd_state_config/data/configuration_templates/wazuh_conf.yaml
