+++ tests/integration/test_agentd/test_agentd_state_config/test_agentd_state_config.py
@@ -60,7 +60,7 @@ from wazuh_testing.tools.monitors.file_monitor import FileMonitor
 from wazuh_testing.utils.configuration import get_test_cases_data
 from wazuh_testing.utils.configuration import load_configuration_template
 from wazuh_testing.utils import callbacks
-from wazuh_testing.utils.services import check_if_process_is_running, control_service
+from wazuh_testing.utils.services import check_if_process_is_running
 
 from . import CONFIGS_PATH, TEST_CASES_PATH
 
@@ -143,5 +143,3 @@ def test_agentd_state_config(test_configuration, test_metadata, remove_state_fil
     wazuh_log_monitor = FileMonitor(WAZUH_LOG_PATH)
     wazuh_log_monitor.start(callback=callbacks.generate_callback(str(test_metadata['event_monitor'])))
     assert (wazuh_log_monitor.callback_result != None), f'Error invalid configuration event not detected'
-
-    time.sleep(3)
